+++ packages/vest/src/core/runtime.ts
@@ -1,14 +1,28 @@
 import { VestRuntime } from 'vest-runtime';
-import { CacheApi, TinyState, cache } from 'vest-utils';
+import {
+  BusType,
+  CacheApi,
+  TinyState,
+  cache,
+  seq,
+  tinyState,
+} from 'vest-utils';
 
 import { Events } from 'BusEvents';
-import { SuiteResult, TFieldName, TGroupName } from 'SuiteResultTypes';
+import {
+  SuiteName,
+  SuiteResult,
+  TFieldName,
+  TGroupName,
+} from 'SuiteResultTypes';
+import { useInitVestBus } from 'VestBus';
 
 export type DoneCallback = (res: SuiteResult<TFieldName, TGroupName>) => void;
 type FieldCallbacks = Record<string, DoneCallbacks>;
 type DoneCallbacks = Array<DoneCallback>;
 
 type StateExtra = {
+  Bus: BusType;
   doneCallbacks: TinyState<DoneCallbacks>;
   fieldCallbacks: TinyState<FieldCallbacks>;
   suiteName: string | undefined;
@@ -17,8 +31,37 @@ type StateExtra = {
 };
 const suiteResultCache = cache<SuiteResult<TFieldName, TGroupName>>();
 
+export function useCreatestate({
+  suiteName,
+}: {
+  suiteName?: SuiteName;
+} = {}): StateExtra {
+  const stateRef: StateExtra = {
+    Bus: useInitVestBus(),
+    doneCallbacks: tinyState.createTinyState<DoneCallbacks>(() => []),
+    fieldCallbacks: tinyState.createTinyState<FieldCallbacks>(() => ({})),
+    suiteId: seq(),
+    suiteName,
+    suiteResultCache,
+  };
+
+  return stateRef;
+}
+
+function useBus() {
+  return useX().Bus;
+}
+
+/*
+  Returns an emitter, but it also has a shortcut for emitting an event immediately
+  by passing an event name.
+*/
+export function useEmit() {
+  return VestRuntime.persist(useBus().emit);
+}
+
 export function usePrepareEmitter<T = void>(event: Events): (arg: T) => void {
-  const emit = VestRuntime.useEmit();
+  const emit = useEmit();
 
   return (arg: T) => emit(event, arg);
 }
